amount received issued fixed in invoices

+++ application/models/pos/M_invoices.php
@@ -1,6 +1,6 @@
 <?php if ( ! defined('BASEPATH')) exit('No direct script access allowed');
 
-class M_sales extends CI_Model{
+class M_invoices extends CI_Model{
     
     public function __construct()
     {
@@ -27,11 +27,11 @@ class M_sales extends CI_Model{
         
         if($sales_id == FALSE)
         {
-            $query = $this->db->get_where('pos_sales',array('company_id'=> $_SESSION['company_id']));
+            $query = $this->db->get_where('pos_invoices',array('company_id'=> $_SESSION['company_id']));
             return $query->result_array();
         }
         
-       $query = $this->db->get_where('pos_sales',array('sale_id'=>$sales_id,'company_id'=> $_SESSION['company_id']));
+       $query = $this->db->get_where('pos_invoices',array('sale_id'=>$sales_id,'company_id'=> $_SESSION['company_id']));
        return $query->result_array();
     }
     
@@ -52,7 +52,7 @@ class M_sales extends CI_Model{
         $this->db->join('pos_customers as c','c.id = s.customer_id','left');
         $this->db->join('pos_employees as e','e.id = s.employee_id','left');
         
-        $query = $this->db->get_where('pos_sales as s',array('s.company_id'=> $_SESSION['company_id']));
+        $query = $this->db->get_where('pos_invoices as s',array('s.company_id'=> $_SESSION['company_id']));
         return $query->result_array();
         
     }
@@ -62,7 +62,7 @@ class M_sales extends CI_Model{
        $this->db->where('total_amount > paid');
        $this->db->where('(total_amount-paid) >',0);
        
-       $query = $this->db->get_where('pos_sales',array('account'=>'credit','register_mode'=>'sale',
+       $query = $this->db->get_where('pos_invoices',array('account'=>'credit','register_mode'=>'sale',
        'customer_id'=>$customer_id,'company_id'=> $_SESSION['company_id']));
        return $query->result_array();
     }
@@ -70,7 +70,7 @@ class M_sales extends CI_Model{
     function updatePaidAmount($invoice_no,$data)
     {
        
-       $this->db->update('pos_sales',$data,array('invoice_no'=>$invoice_no,'company_id'=> $_SESSION['company_id']));
+       $this->db->update('pos_invoices',$data,array('invoice_no'=>$invoice_no,'company_id'=> $_SESSION['company_id']));
        
     }
     
@@ -81,8 +81,8 @@ class M_sales extends CI_Model{
 //       $this->db->select('A.sale_date,A.amount_due,A.register_mode,A.employee_id,A.discount_value,A.customer_id,
 //       A.currency_id,A.description,A.no_of_pkg,A.pkg_desc,A.gd_no,A.gd_date,A.awb_no,A.awb_date,
 //       B.unit,B.item_id,B.size_id,B.color_id,B.item_unit_price,B.item_cost_price,B.quantity_sold,B.discount_percent');
-//       $this->db->join('pos_sales_items as B','A.sale_id = B.sale_id');
-//       $query = $this->db->get_where('pos_sales as A',array('A.invoice_no'=>$new_invoice_no,'A.company_id'=> $_SESSION['company_id']));
+//       $this->db->join('pos_invoices_items as B','A.sale_id = B.sale_id');
+//       $query = $this->db->get_where('pos_invoices as A',array('A.invoice_no'=>$new_invoice_no,'A.company_id'=> $_SESSION['company_id']));
 //       return $query->result_array();
 //       
 //    }
@@ -93,8 +93,8 @@ class M_sales extends CI_Model{
        A.currency_id,A.description,A.invoice_no,A.account,A.is_taxable,A.business_address,A.total_tax,A.total_amount,
        B.unit_id,B.item_id,B.item_unit_price,B.item_cost_price,B.quantity_sold,B.description as item_desc,
        B.discount_percent,B.discount_value,B.tax_rate,B.tax_id,B.inventory_acc_code,B.account_code');
-       $this->db->join('pos_sales_items as B','A.sale_id = B.sale_id');
-       $query = $this->db->get_where('pos_sales as A',array('A.invoice_no'=>$new_invoice_no,'A.company_id'=> $_SESSION['company_id']));
+       $this->db->join('pos_invoices_items as B','A.sale_id = B.sale_id');
+       $query = $this->db->get_where('pos_invoices as A',array('A.invoice_no'=>$new_invoice_no,'A.company_id'=> $_SESSION['company_id']));
        return $query->result_array();
        
     }
@@ -102,7 +102,7 @@ class M_sales extends CI_Model{
     function get_sales_by_invoice($invoice_no)
     {   
         $this->db->where(array('invoice_no'=>$invoice_no,'company_id'=>$_SESSION['company_id']));
-        $query = $this->db->get('pos_sales');
+        $query = $this->db->get('pos_invoices');
         return $query->result_array();
        
     }
@@ -113,10 +113,10 @@ class M_sales extends CI_Model{
     //    A.currency_id,A.description,A.invoice_no,A.account,A.is_taxable,
     //    B.unit_id,B.item_id,B.size_id,B.item_unit_price,B.item_cost_price,B.quantity_sold,B.exchange_rate,B.service,
     //    B.discount_percent,B.discount_value,B.tax_rate,B.tax_id,B.inventory_acc_code');
-    //    $this->db->join('pos_sales_items as B','A.sale_id = B.sale_id');
+    //    $this->db->join('pos_invoices_items as B','A.sale_id = B.sale_id');
        
        $this->db->where(array('invoice_no'=>$invoice_no,'company_id'=>$_SESSION['company_id']));
-       $query = $this->db->get('pos_sales_items');
+       $query = $this->db->get('pos_invoices_items');
        return $query->result_array();
        
     }
@@ -126,7 +126,7 @@ class M_sales extends CI_Model{
         $this->db->order_by('CAST(SUBSTR(invoice_no,2) AS UNSIGNED) DESC');
         $this->db->select('SUBSTR(invoice_no,2) as invoice_no');
         $this->db->where('company_id', $_SESSION['company_id']);
-        $query = $this->db->get('pos_sales',1);
+        $query = $this->db->get('pos_invoices',1);
         return $query->row()->invoice_no;
     }
     
@@ -137,14 +137,14 @@ class M_sales extends CI_Model{
         $this->db->select('invoice_no');
         $this->db->where('company_id', $_SESSION['company_id']);
         $this->db->where('SUBSTR(invoice_no,1,4)','inv-');
-        $query = $this->db->get('pos_sales',1);
+        $query = $this->db->get('pos_invoices',1);
         return $query->row()->invoice_no;
     }
     
     public function get_totalCostBysaleID($invoice_no)
     {
        $this->db->select('SUM(item_unit_price*quantity_sold) as price, SUM(discount_value) as discount_value');   
-       $query = $this->db->get_where('pos_sales_items',array('invoice_no'=>$invoice_no));
+       $query = $this->db->get_where('pos_invoices_items',array('invoice_no'=>$invoice_no));
        $rows = $query->row();
        if($rows)
        {
@@ -171,13 +171,13 @@ class M_sales extends CI_Model{
     }
     
 
-    function delete($invoice_no)
+    function delete($invoice_no,$edit=false)
     {
         $this->db->trans_start();
 
-        $this->db->delete('pos_sales',array('invoice_no'=>$invoice_no,'company_id'=> $_SESSION['company_id']));
+        $this->db->delete('pos_invoices',array('invoice_no'=>$invoice_no,'company_id'=> $_SESSION['company_id']));
         
-        $this->db->delete('pos_sales_items',array('invoice_no'=>$invoice_no,'company_id'=> $_SESSION['company_id']));
+        $this->db->delete('pos_invoices_items',array('invoice_no'=>$invoice_no,'company_id'=> $_SESSION['company_id']));
         
         $sales = $this->get_sales_inv_payment($invoice_no);
         foreach($sales as $key => $list)
@@ -187,7 +187,10 @@ class M_sales extends CI_Model{
             $this->db->delete('acc_entry_items',array('invoice_no'=>$list['invoice_no'],'company_id'=> $_SESSION['company_id']));
 
         }
-        $this->db->delete('pos_sales_inv_payment',array('sales_invoice_no'=>$invoice_no,'company_id'=> $_SESSION['company_id']));
+        if($edit == false)
+        {
+            $this->db->delete('pos_sales_inv_payment',array('sales_invoice_no'=>$invoice_no,'company_id'=> $_SESSION['company_id']));
+        }
 
         $this->db->delete('acc_entries',array('invoice_no'=>$invoice_no,'company_id'=> $_SESSION['company_id']));
         $this->db->delete('acc_entry_items',array('invoice_no'=>$invoice_no,'company_id'=> $_SESSION['company_id']));
@@ -202,8 +205,8 @@ class M_sales extends CI_Model{
     {
         $data = 0;
         $this->db->select('SUM(rt.item_cost_price*rt.quantity_sold) as amount, r.currency_id,it.category_id');   
-        $this->db->join('pos_sales_items as rt','rt.item_id = it.item_id','left');
-        $this->db->join('pos_sales as r','r.sale_id = rt.sale_id','left');
+        $this->db->join('pos_invoices_items as rt','rt.item_id = it.item_id','left');
+        $this->db->join('pos_invoices as r','r.sale_id = rt.sale_id','left');
         $this->db->group_by('it.category_id');
         $query = $this->db->get_where('pos_items it',array('r.company_id'=>$_SESSION['company_id']));
